Deepthi M/Abinaya U| GOK-230 |Refactored folder structure to move stock-receipt inside inventory folder.
